Add measureSamples and estimateInnerLoops
Instead of having the caller having to override exercise to change the
number of loops over run have a new routine, measureSamples.
The method measureSamles has several named parameters of which samples
and minExerciseInMilli are used to by estimateInnerLoops do it to come
up with a value for the loops parameter to exercise.

This make the benchmark much easier to use and will allow more data to
be available for analysis to other programs.